backend/requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.1
cffi==1.14.6
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
defusedxml==0.7.1
drf-yasg
Django==3.2
django-cors-headers
django-environ==0.4.5
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.2
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-extra-fields==3.1.1
flake8==3.9.2
flake8-broken-line==0.3.0
flake8-isort==4.0.0
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
iniconfig==1.1.1
isort==5.8.0
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
more-itertools==8.2.0
oauthlib==3.1.1
packaging==20.3
pep8-naming==0.12.0
Pillow==8.3.1
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.0.1
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
testfixtures==6.18.0
toml==0.10.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0